feat: include global storage equal criterion test
fix: use uuid for null migration path to prevent sub millisecond collisions
fix: use explicit ledger version for global storage criteria
